feat: refactor shared custom button cell
