+++ packages/extension/src/ui/features/accountActivity/AccountActivity.tsx
@@ -1,50 +1,19 @@
-import { FC, Fragment, Suspense, useCallback, useMemo } from "react"
+import { HeaderCell } from "@argent/ui"
+import { FC, Fragment } from "react"
 import { useNavigate } from "react-router-dom"
-import styled from "styled-components"
 
 import { IExplorerTransaction } from "../../../shared/explorer/type"
 import { Token } from "../../../shared/token/type"
-import { useAppState } from "../../app.state"
-import { ErrorBoundary } from "../../components/ErrorBoundary"
-import { ErrorBoundaryFallback } from "../../components/ErrorBoundaryFallback"
-import { Spinner } from "../../components/Spinner"
 import { TransactionStatusIndicator } from "../../components/StatusIndicator"
 import { routes } from "../../routes"
-import { formatDate } from "../../services/dates"
-import { useAspectContractAddresses } from "../accountNfts/aspect.service"
 import { Account } from "../accounts/Account"
-import { useAccountTransactions } from "../accounts/accountTransactions.state"
-import { SectionHeader } from "../accounts/SectionHeader"
-import { useTokensInNetwork } from "../accountTokens/tokens.state"
-import { PendingTransactionsContainer } from "./PendingTransactions"
 import { TransactionListItem } from "./TransactionListItem"
-import { TransactionsListWrapper } from "./TransactionsListWrapper"
 import { transformExplorerTransaction, transformTransaction } from "./transform"
-import {
-  isActivityTransaction,
-  isExplorerTransaction,
-  isVoyagerTransaction,
-} from "./transform/is"
+import { isActivityTransaction, isExplorerTransaction } from "./transform/is"
 import { LoadMoreTrigger } from "./ui/LoadMoreTrigger"
 import { ActivityTransaction } from "./useActivity"
-import { useArgentExplorerAccountTransactionsInfinite } from "./useArgentExplorer"
 
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding-top: 16px;
-  margin-bottom: 68px;
-`
-
-const Header = styled.h2`
-  font-weight: 600;
-  font-size: 32px;
-  line-height: 38.4px;
-  margin-bottom: 25px;
-  text-align: center;
-`
-
-interface IAccountActivity {
+interface AccountActivityProps {
   account: Account
   tokensByNetwork?: Token[]
   nftContractAddresses?: string[]
@@ -53,7 +22,7 @@ interface IAccountActivity {
   onLoadMore: () => void
 }
 
-export const AccountActivity: FC<IAccountActivity> = ({
+export const AccountActivity: FC<AccountActivityProps> = ({
   account,
   tokensByNetwork,
   nftContractAddresses,
@@ -66,226 +35,64 @@ export const AccountActivity: FC<IAccountActivity> = ({
     <>
       {Object.entries(activity).map(([dateLabel, transactions]) => (
         <Fragment key={dateLabel}>
-          <SectionHeader>{dateLabel}</SectionHeader>
-          <TransactionsListWrapper>
-            {transactions.map((transaction) => {
-              if (isActivityTransaction(transaction)) {
-                const { hash, isRejected } = transaction
-                const transactionTransformed = transformTransaction({
-                  transaction,
+          <HeaderCell>{dateLabel}</HeaderCell>
+          {transactions.map((transaction) => {
+            if (isActivityTransaction(transaction)) {
+              const { hash, isRejected } = transaction
+              const transactionTransformed = transformTransaction({
+                transaction,
+                accountAddress: account.address,
+                tokensByNetwork,
+                nftContractAddresses,
+              })
+              if (transactionTransformed) {
+                return (
+                  <TransactionListItem
+                    key={hash}
+                    transactionTransformed={transactionTransformed}
+                    network={account.network}
+                    onClick={() => navigate(routes.transactionDetail(hash))}
+                  >
+                    {isRejected ? (
+                      <div style={{ display: "flex" }}>
+                        <TransactionStatusIndicator color={"red"} />
+                      </div>
+                    ) : null}
+                  </TransactionListItem>
+                )
+              }
+              return null
+            } else if (isExplorerTransaction(transaction)) {
+              const explorerTransactionTransformed =
+                transaction &&
+                transformExplorerTransaction({
+                  explorerTransaction: transaction,
                   accountAddress: account.address,
                   tokensByNetwork,
                   nftContractAddresses,
                 })
-                if (transactionTransformed) {
-                  return (
+              if (explorerTransactionTransformed) {
+                const { transactionHash } = transaction
+                const loadMore = loadMoreHashes.includes(transactionHash)
+                return (
+                  <Fragment key={transactionHash}>
                     <TransactionListItem
-                      key={hash}
-                      transactionTransformed={transactionTransformed}
+                      transactionTransformed={explorerTransactionTransformed}
                       network={account.network}
-                      onClick={() => navigate(routes.transactionDetail(hash))}
-                    >
-                      {isRejected ? (
-                        <div style={{ display: "flex" }}>
-                          <TransactionStatusIndicator color={"red"} />
-                        </div>
-                      ) : null}
-                    </TransactionListItem>
-                  )
-                }
-                return null
-              } else if (isExplorerTransaction(transaction)) {
-                const explorerTransactionTransformed =
-                  transaction &&
-                  transformExplorerTransaction({
-                    explorerTransaction: transaction,
-                    accountAddress: account.address,
-                    tokensByNetwork,
-                    nftContractAddresses,
-                  })
-                if (explorerTransactionTransformed) {
-                  const { transactionHash } = transaction
-                  const loadMore = loadMoreHashes.includes(transactionHash)
-                  return (
-                    <Fragment key={transactionHash}>
-                      <TransactionListItem
-                        transactionTransformed={explorerTransactionTransformed}
-                        network={account.network}
-                        onClick={() =>
-                          navigate(routes.transactionDetail(transactionHash))
-                        }
-                      />
-                      {loadMore && <LoadMoreTrigger onLoadMore={onLoadMore} />}
-                    </Fragment>
-                  )
-                }
-              } else {
-                return null
+                      onClick={() =>
+                        navigate(routes.transactionDetail(transactionHash))
+                      }
+                    />
+                    {loadMore && <LoadMoreTrigger onLoadMore={onLoadMore} />}
+                  </Fragment>
+                )
               }
-            })}
-          </TransactionsListWrapper>
+            } else {
+              return null
+            }
+          })}
         </Fragment>
       ))}
     </>
-  )
-}
-
-interface IAccountActivityContainer {
-  account: Account
-}
-
-const PAGE_SIZE = 10
-
-export const AccountActivityContainer: FC<IAccountActivityContainer> = ({
-  account,
-}) => {
-  const { switcherNetworkId } = useAppState()
-  const tokensByNetwork = useTokensInNetwork(switcherNetworkId)
-  const { data: nftContractAddresses } = useAspectContractAddresses()
-  const { data, setSize } = useArgentExplorerAccountTransactionsInfinite({
-    accountAddress: account.address,
-    network: switcherNetworkId,
-    pageSize: PAGE_SIZE,
-  })
-
-  const explorerTransactions = useMemo(() => {
-    if (!data) {
-      return
-    }
-    return data.flat()
-  }, [data])
-
-  const isEmpty = data?.[0]?.length === 0
-  const isReachingEnd =
-    isEmpty || (data && data[data.length - 1]?.length < PAGE_SIZE)
-
-  const { transactions } = useAccountTransactions(account)
-  const voyagerTransactions = useMemo(() => {
-    // RECEIVED transactions are already shown as pending
-    return transactions.filter(
-      (transaction) =>
-        transaction.status !== "RECEIVED" && !transaction.meta?.isDeployAccount,
-    )
-  }, [transactions])
-  const mergedTransactions = useMemo(() => {
-    if (!explorerTransactions) {
-      return {
-        transactions: voyagerTransactions,
-      }
-    }
-    const matchedHashes: string[] = []
-
-    const mergedTransactions = voyagerTransactions.map((voyagerTransaction) => {
-      const explorerTransaction = explorerTransactions.find(
-        (explorerTransaction) =>
-          explorerTransaction.transactionHash === voyagerTransaction.hash,
-      )
-      if (explorerTransaction) {
-        if (!explorerTransaction.timestamp) {
-          explorerTransaction.timestamp = voyagerTransaction.timestamp
-        }
-        matchedHashes.push(voyagerTransaction.hash)
-        return explorerTransaction
-      }
-      return voyagerTransaction
-    })
-
-    const unmatchedExplorerTransactions = explorerTransactions.filter(
-      (explorerTransaction) =>
-        !matchedHashes.includes(explorerTransaction.transactionHash),
-    )
-
-    const transactionsWithoutTimestamp = []
-    for (const transaction of unmatchedExplorerTransactions) {
-      if (transaction.timestamp) {
-        mergedTransactions.push(transaction)
-      } else {
-        transactionsWithoutTimestamp.push(transaction)
-      }
-    }
-
-    const sortedTransactions = mergedTransactions.sort(
-      (a, b) => b.timestamp - a.timestamp,
-    )
-
-    return {
-      transactions: sortedTransactions,
-      transactionsWithoutTimestamp,
-    }
-  }, [explorerTransactions, voyagerTransactions])
-
-  const { mergedActivity, loadMoreHashes } = useMemo(() => {
-    const mergedActivity: Record<
-      string,
-      Array<ActivityTransaction | IExplorerTransaction>
-    > = {}
-    const { transactions, transactionsWithoutTimestamp } = mergedTransactions
-    let lastExplorerTransactionHash
-    for (const transaction of transactions) {
-      const date = new Date(transaction.timestamp * 1000).toISOString()
-      const dateLabel = formatDate(date)
-      mergedActivity[dateLabel] ||= []
-      if (isVoyagerTransaction(transaction)) {
-        const { hash, meta, status } = transaction
-        const isRejected = status === "REJECTED"
-        const activityTransaction: ActivityTransaction = {
-          hash,
-          date,
-          meta,
-          isRejected,
-        }
-        mergedActivity[dateLabel].push(activityTransaction)
-      } else {
-        mergedActivity[dateLabel].push(transaction)
-        lastExplorerTransactionHash = transaction.transactionHash
-      }
-    }
-
-    const loadMoreHashes = []
-
-    if (lastExplorerTransactionHash) {
-      loadMoreHashes.push(lastExplorerTransactionHash)
-    }
-
-    if (transactionsWithoutTimestamp && transactionsWithoutTimestamp.length) {
-      mergedActivity["Unknown date"] = transactionsWithoutTimestamp
-      loadMoreHashes.push(
-        transactionsWithoutTimestamp[transactionsWithoutTimestamp.length - 1]
-          .transactionHash,
-      )
-    }
-    return {
-      mergedActivity,
-      loadMoreHashes,
-    }
-  }, [mergedTransactions])
-
-  const onLoadMore = useCallback(() => {
-    if (!isReachingEnd) {
-      setSize((size) => size + 1)
-    }
-  }, [isReachingEnd, setSize])
-
-  return (
-    <Container>
-      <Header>Activity</Header>
-      <PendingTransactionsContainer account={account} />
-      <ErrorBoundary
-        fallback={
-          <ErrorBoundaryFallback title="Seems like Voyager API is down..." />
-        }
-      >
-        <Suspense fallback={<Spinner size={64} style={{ marginTop: 40 }} />}>
-          <AccountActivity
-            activity={mergedActivity}
-            loadMoreHashes={loadMoreHashes}
-            account={account}
-            tokensByNetwork={tokensByNetwork}
-            nftContractAddresses={nftContractAddresses}
-            onLoadMore={onLoadMore}
-          />
-        </Suspense>
-      </ErrorBoundary>
-    </Container>
   )
 }
