+++ packages/extension/src/ui/features/accountActivity/TransactionListItem.tsx
@@ -1,14 +1,13 @@
 import { FC, ReactNode, useMemo } from "react"
-import styled, { css } from "styled-components"
+import styled from "styled-components"
 
 import { Network } from "../../../shared/network"
-import { makeClickable } from "../../services/a11y"
+import { CustomButtonCell } from "../../components/CustomButtonCell"
 import { PrettyAccountAddress } from "../accounts/PrettyAccountAddress"
 import {
   TokenDetailsWrapper,
   TokenTextGroup,
   TokenTitle,
-  TokenWrapper,
 } from "../accountTokens/TokenListItemDeprecated"
 import {
   isDeclareContractTransaction,
@@ -24,27 +23,6 @@ import { NFTAccessory } from "./ui/NFTAccessory"
 import { SwapAccessory } from "./ui/SwapAccessory"
 import { TransactionIcon } from "./ui/TransactionIcon"
 import { TransferAccessory } from "./ui/TransferAccessory"
-
-export const TransactionsListWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-`
-
-const Container = styled(TokenWrapper)<{
-  highlighted?: boolean
-}>`
-  cursor: pointer;
-
-  ${({ highlighted }) =>
-    highlighted &&
-    css`
-      background-color: rgba(255, 255, 255, 0.1);
-    `}
-
-  &:hover, &:focus {
-    background-color: rgba(255, 255, 255, 0.15);
-  }
-`
 
 const TransactionSubtitle = styled.div`
   font-size: 13px;
@@ -64,7 +42,7 @@ const TitleAddressPrefix = styled.div`
 
 const TitleAddress = styled.div``
 
-export interface ITransactionListItem {
+export interface TransactionListItemProps {
   transactionTransformed: TransformedTransaction
   network: Network
   highlighted?: boolean
@@ -72,16 +50,14 @@ export interface ITransactionListItem {
   children?: ReactNode | ReactNode[]
 }
 
-export const TransactionListItem: FC<ITransactionListItem> = ({
+export const TransactionListItem: FC<TransactionListItemProps> = ({
   transactionTransformed,
   network,
   highlighted,
-  onClick,
   children,
   ...props
 }) => {
   const { action, displayName, dapp } = transactionTransformed
-  console.log(transactionTransformed)
   const isNFT = isNFTTransaction(transactionTransformed)
   const isNFTTransfer = isNFTTransferTransaction(transactionTransformed)
   const isTransfer = isTokenTransferTransaction(transactionTransformed)
@@ -155,7 +131,7 @@ export const TransactionListItem: FC<ITransactionListItem> = ({
   ])
 
   return (
-    <Container {...makeClickable(onClick)} highlighted={highlighted} {...props}>
+    <CustomButtonCell highlighted={highlighted} {...props}>
       <TransactionIcon transaction={transactionTransformed} size={40} />
       <TokenDetailsWrapper>
         <TokenTextGroup>
@@ -165,6 +141,6 @@ export const TransactionListItem: FC<ITransactionListItem> = ({
       </TokenDetailsWrapper>
       {accessory}
       {children}
-    </Container>
+    </CustomButtonCell>
   )
 }
